+++ src/apis/endpoints.ts
@@ -1,0 +1,4 @@
+export const enum APIEndPoints  {
+    USERS = " https://gorest.co.in/public/v2/users",
+    POSTS =  "https://gorest.co.in/public/v2/posts",
+}
