+++ src/components/UsersList.tsx
@@ -1,4 +1,4 @@
-import { useUsers } from '../hooks/FetchUsers';
+import { useUsers } from '../apis/FetchUsers';
 import UserDetails from './UserDetails';
 
 const UserList = () => {
@@ -17,7 +17,7 @@ const UserList = () => {
         textAlign: 'justify',
         marginLeft: '200px',
         gap: '20px',
-        marginTop:'50px'
+        marginTop: '50px',
       }}
     >
       {data?.map(user => (
